Migrate doctor_validator to null safety (#79682)
