+++ packages/flutter_tools/lib/src/android/android_studio_validator.dart
@@ -11,7 +11,7 @@ import '../base/file_system.dart';
 import '../base/platform.dart';
 import '../base/user_messages.dart';
 import '../base/version.dart';
-import '../doctor.dart';
+import '../doctor_validator.dart';
 import '../intellij/intellij.dart';
 import 'android_studio.dart';
