+++ packages/flutter_tools/lib/src/macos/cocoapods_validator.dart
@@ -5,7 +5,7 @@
 // @dart = 2.8
 
 import '../base/user_messages.dart';
-import '../doctor.dart';
+import '../doctor_validator.dart';
 import 'cocoapods.dart';
 
 /// A validator that confirms cocoapods is in a valid state.
