+++ packages/flutter_tools/lib/src/vscode/vscode.dart
@@ -11,7 +11,7 @@ import '../base/platform.dart';
 import '../base/utils.dart';
 import '../base/version.dart';
 import '../convert.dart';
-import '../doctor.dart';
+import '../doctor_validator.dart';
 
 // Include VS Code insiders (useful for debugging).
 const bool _includeInsiders = false;
