+++ packages/flutter_tools/lib/src/vscode/vscode_validator.dart
@@ -8,7 +8,7 @@ import '../base/file_system.dart';
 import '../base/platform.dart';
 import '../base/user_messages.dart';
 import '../base/version.dart';
-import '../doctor.dart';
+import '../doctor_validator.dart';
 import 'vscode.dart';
 
 class VsCodeValidator extends DoctorValidator {
