+++ packages/flutter_tools/lib/src/web/workflow.dart
@@ -7,7 +7,7 @@
 import 'package:meta/meta.dart';
 
 import '../base/platform.dart';
-import '../doctor.dart';
+import '../doctor_validator.dart';
 import '../features.dart';
 
 class WebWorkflow extends Workflow {
