+++ packages/flutter_tools/test/general.shard/intellij/intellij_test.dart
@@ -10,7 +10,7 @@ import 'package:archive/archive.dart';
 import 'package:file/file.dart';
 import 'package:file/memory.dart';
 import 'package:flutter_tools/src/base/file_system.dart';
-import 'package:flutter_tools/src/doctor.dart';
+import 'package:flutter_tools/src/doctor_validator.dart';
 import 'package:flutter_tools/src/intellij/intellij.dart';
 
 import '../../src/common.dart';
