+++ packages/flutter_tools/test/general.shard/macos/xcode_validator_test.dart
@@ -6,7 +6,7 @@
 
 import 'package:flutter_tools/src/base/user_messages.dart';
 import 'package:flutter_tools/src/base/version.dart';
-import 'package:flutter_tools/src/doctor.dart';
+import 'package:flutter_tools/src/doctor_validator.dart';
 import 'package:flutter_tools/src/macos/xcode.dart';
 import 'package:flutter_tools/src/macos/xcode_validator.dart';
 import 'package:mockito/mockito.dart';
